ia-service/requirements.txt
fastapi
uvicorn
google-generativeai
supabase
